# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Documentation API MYAPPS LOGIN",
        "description": "Ini Dokumentasi API untuk semua respon yang didapatkan dari API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:8089",
    "basePath": "/api/",
    "tags": [{
        "name": "API MYAPPS LOGIN",
        "description": "Get Data API MYAPPS LOGIN"
    }],
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/login": {
            "parameters": [
                {
                    "name": "nik",
                    "in": "path",
                    "required": true,
                    "description": "NIK",
                    "type": "string"
                },
                {
                    "name": "password",
                    "in": "path",
                    "required": true,
                    "description": "Password",
                    "type": "string"
                }
            ],
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Post Login",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/menuprogram": {
            "parameters": [{
                "name": "Authorization",
                "in": "header",
                "required": true,
                "description": "Token Authorization",
                "type": "string"
            }],
            "get": {
                "tags": [
                    "Menu Program"
                ],
                "summary": "Get Menu Program",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/list-menu": {
            "parameters": [{
                "name": "Authorization",
                "in": "header",
                "required": true,
                "description": "Token Authorization",
                "type": "string"
            }],
            "get": {
                "tags": [
                    "List Menu"
                ],
                "summary": "Get List Menu",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/user-data": {
            "parameters": [{
                "name": "Authorization",
                "in": "header",
                "required": true,
                "description": "Token Authorization",
                "type": "string"
            }],
            "get": {
                "tags": [
                    "User Data"
                ],
                "summary": "Get User Data",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/reset-password": {
            "parameters": [
                {
                    "name": "Authorization",
                    "in": "header",
                    "required": true,
                    "description": "Token Authorization",
                    "type": "string"
                },
                {
                    "name": "password",
                    "in": "path",
                    "required": true,
                    "description": "Password",
                    "type": "string"
                }
            ],
            "post": {
                "tags": [
                    "Reset Password"
                ],
                "summary": "Post Reset Password",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/ismenu-exist": {
            "parameters": [
                {
                    "name": "Authorization",
                    "in": "header",
                    "required": true,
                    "description": "Token Authorization",
                    "type": "string"
                },
                {
                    "name": "code",
                    "in": "path",
                    "required": true,
                    "description": "Kode Menu",
                    "type": "string"
                }
            ],
            "post": {
                "tags": [
                    "Menu"
                ],
                "summary": "Post Exist Menu",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/delete-menu": {
            "parameters": [
                {
                    "name": "Authorization",
                    "in": "header",
                    "required": true,
                    "description": "Token Authorization",
                    "type": "string"
                },
                {
                    "name": "code",
                    "in": "path",
                    "required": true,
                    "description": "Kode Menu",
                    "type": "string"
                }
            ],
            "delete": {
                "tags": [
                    "Menu"
                ],
                "summary": "Delete Menu",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/update-menu": {
            "parameters": [
                {
                    "name": "Authorization",
                    "in": "header",
                    "required": true,
                    "description": "Token Authorization",
                    "type": "string"
                },
                {
                    "name": "code",
                    "in": "path",
                    "required": true,
                    "description": "Kode Menu",
                    "type": "string"
                },
                {
                    "name": "name",
                    "in": "path",
                    "required": true,
                    "description": "Name Menu",
                    "type": "string"
                },
                {
                    "name": "file",
                    "in": "path",
                    "required": true,
                    "description": "File Menu",
                    "type": "string"
                },
                {
                    "name": "folder",
                    "in": "path",
                    "required": true,
                    "description": "Folder Menu",
                    "type": "string"
                }
            ],
            "post": {
                "tags": [
                    "Menu"
                ],
                "summary": "Update Menu",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/add-menu": {
            "parameters": [
                {
                    "name": "Authorization",
                    "in": "header",
                    "required": true,
                    "description": "Token Authorization",
                    "type": "string"
                },
                {
                    "name": "code",
                    "in": "path",
                    "required": true,
                    "description": "Kode Menu",
                    "type": "string"
                },
                {
                    "name": "name",
                    "in": "path",
                    "required": true,
                    "description": "Name Menu",
                    "type": "string"
                },
                {
                    "name": "file",
                    "in": "path",
                    "required": true,
                    "description": "File Menu",
                    "type": "string"
                },
                {
                    "name": "folder",
                    "in": "path",
                    "required": true,
                    "description": "Folder Menu",
                    "type": "string"
                }
            ],
            "post": {
                "tags": [
                    "Menu"
                ],
                "summary": "Add Menu",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/forgot-pass/{userlogin}": {
            "parameters": [
                {
                    "name": "userlogin",
                    "in": "path",
                    "required": true,
                    "description": "User Login",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Forgot Password"
                ],
                "summary": "Get Forgot Password",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/get-email/{userlogin}": {
            "parameters": [
                {
                    "name": "userlogin",
                    "in": "path",
                    "required": true,
                    "description": "User Login",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Email"
                ],
                "summary": "Get Email",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/menu/{kodeprogram}": {
            "parameters": [
                {
                    "name": "Authorization",
                    "in": "header",
                    "required": true,
                    "description": "Token Authorization",
                    "type": "string"
                },
                {
                    "name": "kodeprogram",
                    "in": "path",
                    "required": true,
                    "description": "Kode Program",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Get Menu",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/submenu/{kodeprogram}/{kodemenu}": {
            "parameters": [
                {
                    "name": "Authorization",
                    "in": "header",
                    "required": true,
                    "description": "Token Authorization",
                    "type": "string"
                },
                {
                    "name": "kodeprogram",
                    "in": "path",
                    "required": true,
                    "description": "Kode Program",
                    "type": "string"
                },                
                {
                    "name": "kodemenu",
                    "in": "path",
                    "required": true,
                    "description": "Kode Menu",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Sub Menu"
                ],
                "summary": "Get Sub Menu",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/logindata/{userlogin}/{password}": {
            "parameters": [
                {
                    "name": "Authorization",
                    "in": "header",
                    "required": true,
                    "description": "Token Authorization",
                    "type": "string"
                },
                {
                    "name": "userlogin",
                    "in": "path",
                    "required": true,
                    "description": "User Login",
                    "type": "string"
                },                
                {
                    "name": "password",
                    "in": "path",
                    "required": true,
                    "description": "Password",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Login"
                ],
                "summary": "Get Login",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/login/{userlogin}/{password}": {
            "parameters": [
                {
                    "name": "Authorization",
                    "in": "header",
                    "required": true,
                    "description": "Token Authorization",
                    "type": "string"
                },
                {
                    "name": "userlogin",
                    "in": "path",
                    "required": true,
                    "description": "User Login",
                    "type": "string"
                },                
                {
                    "name": "password",
                    "in": "path",
                    "required": true,
                    "description": "Password",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Login"
                ],
                "summary": "Get Login",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    }
}